add training mode
